# Fix type instability in normals