--- a/src/geometry_primitives.jl
+++ b/src/geometry_primitives.jl
@@ -107,7 +107,12 @@ Compute all vertex normals.
 """
 function normals(vertices::AbstractVector{<:AbstractPoint{3,T}}, faces::AbstractVector{F};
                  normaltype=Vec{3,T}) where {T,F<:NgonFace}
-    normals_result = zeros(normaltype, length(vertices)) # initilize with same type as verts but with 0
+    return normals(vertices, faces, normaltype)
+end
+
+function normals(vertices::AbstractVector{<:AbstractPoint{3,T}}, faces::AbstractVector{F},
+                 ::Type{N}) where {T,F<:NgonFace,N}
+    normals_result = zeros(N, length(vertices)) # initilize with same type as verts but with 0
     for face in faces
         v = metafree.(vertices[face])
         # we can get away with two edges since faces are planar.